fix(win): resolve Windows ARM compilation and FS test issues

+++ native/include/fs_native.h
@@ -16,19 +16,20 @@ typedef void (*FSEventCallback)(const char* path, int type, int isDir);
 
 /**
  * Declares file system watching system calls.
+ * Uses double for handles to avoid truncation on 64-bit platforms in Haxe.
  */
 extern "C" {
     int fs_watch_start(const char* path, FSEventCallback callback);
     void fs_watch_stop_all();
 
-    int fs_file_lock(const char* path, int exclusive, int wait);
-    void fs_file_unlock(int id);
+    double fs_file_lock(const char* path, int exclusive, int wait);
+    void fs_file_unlock(double id);
 
-    int fs_mmap_open(const char* path, int size, int writable);
-    void fs_mmap_close(int id);
-    int fs_mmap_read(int id, int offset, char* buffer, int length);
-    int fs_mmap_write(int id, int offset, const char* buffer, int length);
-    void fs_mmap_flush(int id);
+    double fs_mmap_open(const char* path, int size, int writable);
+    void fs_mmap_close(double id);
+    int fs_mmap_read(double id, int offset, char* buffer, int length);
+    int fs_mmap_write(double id, int offset, const char* buffer, int length);
+    void fs_mmap_flush(double id);
 
     int fs_set_xattr(const char* path, const char* name, const unsigned char* value, int value_len);
     int fs_get_xattr(const char* path, const char* name, unsigned char* buffer, int buffer_len);
